feat: add permission checks for notifications
This commit includes updates to `HomeView.kt`, `AttendanceWorker.kt`, and `MainActivity.kt`. It introduces permission checks for posting notifications in the AttendanceWorker class. If the requisite permission is not granted, the application now sends an intent to request that permission. This ensures the application adheres to Android's security best practices.

Additionally, the Github workflow for generating the debug pre-release APK now runs on MacOS instead of Ubuntu.