feat: replace BroadcastReceiver with IntentService in AttendanceWorker
This commit replaces the BroadcastReceiver with an IntentService in the AttendanceWorker class. It also updates the PendingIntent to use getService instead of getBroadcast. A new file for AttendanceService has been created, and the attendance mark is now being handled inside this service. Furthermore, the time interval for the AttendanceWorker has been changed from 1 minute to 3 minutes.