feat: add "Reglab" button and corresponding view
This commit introduces a new button "Reglab" in the HomeView screen. Clicking on it will navigate to a new view named ReglabView. The new view is also added to the Screen enum to handle navigation. Lastly, the logic to display the ReglabView based on the current screen state is added in MainActivity. Also, data classes for the Reglab feature are introduced in MainViewModel.